extract Delete Confirmation Modal and related logic in a component
call the component from `InfrastructureMappingsList`
