+++ app/javascript/react/screens/App/Overview/Overview.js
@@ -1,14 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Card,
-  Breadcrumb,
-  CardGrid,
-  Spinner,
-  Button,
-  Icon,
-  Modal
-} from 'patternfly-react';
+import { Card, Breadcrumb, CardGrid, Spinner } from 'patternfly-react';
 import Toolbar from '../../../config/Toolbar';
 import * as AggregateCards from './components/AggregateCards';
 import InfrastructureMappingsList from './components/InfrastructureMappingsList/InfrastructureMappingsList';
@@ -191,52 +183,6 @@ class Overview extends React.Component {
     history.push(path);
   };
 
-  unUsedMappingInPlans = mappingToDelete => {
-    const planNames = [];
-
-    const {
-      notStartedTransformationPlans,
-      finishedWithErrorTransformationPlans
-    } = this.props;
-
-    const plansWithMappingToBeDeleted = notStartedTransformationPlans.filter(
-      plan =>
-        plan.options.config_info.transformation_mapping_id ===
-        mappingToDelete.id
-    );
-    plansWithMappingToBeDeleted.map(plan => planNames.push(plan.name));
-
-    const plansWithErrorWithMappingToBeDeleted = finishedWithErrorTransformationPlans.filter(
-      plan =>
-        plan.options.config_info.transformation_mapping_id ===
-        mappingToDelete.id
-    );
-    plansWithErrorWithMappingToBeDeleted.map(plan => planNames.push(plan.name));
-
-    if (planNames.length > 0) {
-      const deleteMessageAboutUnMigratedVMs = __(
-        'The infrastructure mapping is associated with migration plans that include unmigrated VMs. Deleting the mapping will prevent you from migrating the VMs in these plans:'
-      );
-      const deleteMessageAboutPlansUsingMapping = (
-        <div>
-          <h4>{deleteMessageAboutUnMigratedVMs}</h4>
-          <strong>{planNames.map(plan => <ul key={plan}>{plan}</ul>)}</strong>
-        </div>
-      );
-      return <div>{deleteMessageAboutPlansUsingMapping}</div>;
-    }
-    return '';
-  };
-
-  displayDeleteMessage = mappingToDelete => {
-    const mappingNameStyled = `<strong>${mappingToDelete.name}</strong>`;
-    const regularDeleteMessage = sprintf(
-      __('Are you sure you want to delete the infrastructure mapping %s ?'),
-      mappingNameStyled
-    );
-    return <div dangerouslySetInnerHTML={{ __html: regularDeleteMessage }} />;
-  };
-
   render() {
     const {
       showMappingWizardAction,
@@ -254,6 +200,7 @@ class Overview extends React.Component {
       notStartedTransformationPlans,
       activeTransformationPlans,
       finishedTransformationPlans,
+      finishedWithErrorTransformationPlans,
       isCreatingTransformationPlanRequest,
       clusters,
       migrationsFilter,
@@ -263,7 +210,8 @@ class Overview extends React.Component {
       hideDeleteConfirmationModalAction,
       setMappingToDeleteAction,
       mappingToDelete,
-      yesToDeleteInfrastructureMappingAction
+      yesToDeleteInfrastructureMappingAction,
+      deleteInfrastructureMappingAction
     } = this.props;
 
     const inProgressRequestsTransformationMappings = () => {
@@ -362,46 +310,22 @@ class Overview extends React.Component {
               showDeleteConfirmationModalAction
             }
             setMappingToDeleteAction={setMappingToDeleteAction}
+            showDeleteConfirmationModal={showDeleteConfirmationModal}
+            hideDeleteConfirmationModalAction={
+              hideDeleteConfirmationModalAction
+            }
+            mappingToDelete={mappingToDelete}
+            yesToDeleteInfrastructureMappingAction={
+              yesToDeleteInfrastructureMappingAction
+            }
+            notStartedTransformationPlans={notStartedTransformationPlans}
+            finishedWithErrorTransformationPlans={
+              finishedWithErrorTransformationPlans
+            }
+            deleteInfrastructureMappingAction={
+              deleteInfrastructureMappingAction
+            }
           />
-          <Modal
-            show={showDeleteConfirmationModal}
-            onHide={hideDeleteConfirmationModalAction}
-          >
-            <Modal.Header>
-              <Modal.CloseButton onClick={hideDeleteConfirmationModalAction} />
-              <Modal.Title>{__('Delete Infrastructure Mapping')}</Modal.Title>
-            </Modal.Header>
-            <Modal.Body className="warning-modal-body">
-              <div className="warning-modal-body--icon">
-                <Icon type="pf" name="delete" />
-              </div>
-              <div className="warning-modal-body--list">
-                <h4>
-                  {mappingToDelete &&
-                    this.unUsedMappingInPlans(mappingToDelete)}
-                </h4>
-                <h4>
-                  {mappingToDelete &&
-                    this.displayDeleteMessage(mappingToDelete)}
-                </h4>
-              </div>
-            </Modal.Body>
-            <Modal.Footer>
-              <Button
-                bsStyle="default"
-                className="btn-cancel"
-                onClick={hideDeleteConfirmationModalAction}
-              >
-                {__('Cancel')}
-              </Button>
-              <Button
-                bsStyle="primary"
-                onClick={yesToDeleteInfrastructureMappingAction}
-              >
-                {__('Delete')}
-              </Button>
-            </Modal.Footer>
-          </Modal>
         </Spinner>
       </div>
     );
